requirements.txt
selenium==4.15.2
fake_useragent==1.4.0
python-dotenv==1.0.0